requirements.txt
blinker==1.6.2
cadquery==2.3.1
cadquery-ocp==7.7.1
casadi==3.6.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
ezdxf==1.0.3
Flask==2.3.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multimethod==1.9.1
nlopt==2.7.1
nptyping==2.0.1
numpy==1.25.2
path==16.7.1
psycopg2-binary==2.9.7
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
typish==1.9.3
urllib3==2.0.4
uvicorn==0.23.2
Werkzeug==2.3.7
